ThirdChapter

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -32,6 +32,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if(app()->isLocal()){
+            Schema::dropIfExists('users');
+        }
     }
 };
