+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('failed_jobs');
+        if(app()->isLocal()){
+            Schema::dropIfExists('failed_jobs');
+        }
     }
 };
