+++ database/migrations/2022_10_11_111955_create_categories_table.php
@@ -33,6 +33,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('categories');
+        if(app()->isLocal()){
+            Schema::dropIfExists('categories');
+        }
     }
 };
