+++ database/migrations/2022_10_11_112020_create_brands_table.php
@@ -33,6 +33,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('brands');
+        if(app()->isLocal()){
+            Schema::dropIfExists('brands');
+        }
     }
 };
