+++ database/migrations/2022_10_11_112025_create_products_table.php
@@ -1,8 +1,11 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use App\Models\Brand;
+use App\Models\Product;
+use App\Models\Category;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
@@ -21,20 +24,28 @@ return new class extends Migration
             $table->text('images')->nullable();
             $table->double('price', 8, 2);
             $table->text('description')->nullable();
-            $table->foreignId('category_id')
+
+            $table->foreignIdFor(Brand::class)
                     ->nullable()
                     ->constrained()
                     ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->foreignId('brand_id')
-                    ->nullable()
-                    ->constrained()
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
+                    ->nullOnDelete('cascade');
             $table->boolean('status')->default(true);
             $table->integer('sort')->default(500);
             $table->timestamps();
             $table->softDeletes();
+        });
+
+        Schema::create('category_product', function (Blueprint $table) {
+            $table->id();
+            $table->foreignIdFor(Category::class)
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            $table->foreignIdFor(Product::class)
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
         });
     }
 
@@ -45,6 +56,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('products');
+        if(app()->isLocal()){
+            Schema::dropIfExists('category_product');
+            Schema::dropIfExists('products');
+        }
     }
 };
