+++ database/migrations/2022_10_11_124138_create_product_raitings_table.php
@@ -37,6 +37,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('product_raitings');
+        if(app()->isLocal()){
+            Schema::dropIfExists('product_raitings');
+        }
     }
 };
