+++ database/migrations/2022_10_11_132602_add_admin_column_to_users_table.php
@@ -25,8 +25,10 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_admin');
-        });
+        if(app()->isLocal()){
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('is_admin');
+            });
+        }
     }
 };
